Use is_autheticated attribute, drop Django < 1.11
Django 1.11 starts the transition from the method
request.user.is_authenticated() to the property
request.user.is_authenticated. Either can be used, but a
RemovedInDjango20Warning is raised if the method is used.

Simplify the code to expect a boolean user.is_authenticated property,
and drop support for the non-supported Django versions 1.8, 1.9, and
1.10.