Add RATELIMIT_IP_META_KEY setting
The RATELIMIT_IP_META_KEY setting allows the application to control the
source of truth for the client IP address for use with `key='ip'` or
`key='user_or_ip'`. It enables an application to set this in a single
place rather than setting `key='header:x-client-ip'`, which means this
can be controlled for local development. Like other parameters in
django_ratelimit, this setting accepts callables and paths to callables.