Add clear exception if REMOTE_ADDR is empty
Certain configurations of Django when running behind a reverse proxy
that communicates with the app server over Unix sockets can lead to
REMOTE_ADDR being empty or missing. In these situations, IP-based
ratelimiting can behave unexpectedly or fail in unclear ways. This makes
the exception for an empty REMOTE_ADDR more clear and points to the
documentation.