Adjust the examples to have actual side-effects
Most of the examples did not any effect on the request when they should have been rate limited. This was initially deceptive as it was not clear that nothing would happen.
These changes show the ways in which you can have a rate limit actually happen.

I would propose that the default behavior would have the `block` option set to True but I would not do that without a larger version number change.